+++ src/components/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
-import ChatPage from '../containers/ChatPage';
-import WelcomePage from '../containers/WelcomePage';
-import configureStore from '../store';
+import ChatPage from './ChatPage';
+import WelcomePage from '../redux/containers/WelcomePage';
+import configureStore from '../redux/store';
 
 const store = configureStore();
