+++ src/components/ChatHeader.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import classNames from 'classnames';
 import { withStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
@@ -8,18 +7,17 @@ import Typography from '@material-ui/core/Typography';
 const styles = theme => ({
   appBar: {
     width: `calc(100% - ${320}px)`,
-  }
+  },
 });
 
 const ChatHeader = ({ classes }) => (
-  <AppBar position="absolute" className={classes.appBar}>
+  <AppBar position='absolute' className={classes.appBar}>
     <Toolbar>
-      <Typography variant="title" color="inherit" noWrap>
+      <Typography variant='title' color='inherit' noWrap>
         DogeCodes React chat
       </Typography>
     </Toolbar>
   </AppBar>
 );
-
 
 export default withStyles(styles)(ChatHeader);
