--- src/containers/ChatPage.js
@@ -1,11 +1,0 @@
-import { connect } from 'react-redux';
-import ChatPage from '../components/ChatPage';
-
-const mapStateToProps = state => ({});
-
-const mapDispatchToProps = dispatch => ({});
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(ChatPage);
