+++ src/pages/chatPage/sideBar/ChatList.jsx
@@ -4,21 +4,21 @@ import List from '@material-ui/core/List';
 import ChatListItem from './ChatListItem';
 
 const styles = theme => ({
-    chatsList: {
-        height: 'calc(100% - 56px)',
-        overflowY: 'scroll',
-    },
+  chatsList: {
+    height: 'calc(100% - 56px)',
+    overflowY: 'scroll',
+  },
 });
 
 const ChatList = ({ classes, chatList }) => (
-    <List className={classes.chatsList}>
-        {chatList.map((chat, index) => (
-            <ChatListItem
-                key={index}
-                chat={chat}
-            />
-        ))}
-    </List>
+  <List className={classes.chatsList}>
+    {chatList.map((chat, index) => (
+      <ChatListItem
+        key={index}
+        chat={chat}
+      />
+    ))}
+  </List>
 );
 
 export default withStyles(styles)(ChatList);
