+++ src/redux/actions/auth.js
@@ -2,11 +2,10 @@ import {
   SIGNUP_REQUEST, SIGNUP_SUCCESS, SIGNUP_FAILURE,
   LOGIN_REQUEST, LOGIN_SUCCESS, LOGIN_FAILURE,
   LOGOUT_REQUEST, LOGOUT_SUCCESS, LOGOUT_FAILURE,
-} from '../constants';
+} from '../constants/index';
 import fetch from 'isomorphic-fetch';
 
 export function signup(username, password) {
-  console.log(username, password);
   return (dispatch) => {
     dispatch({
       type: SIGNUP_REQUEST,
@@ -19,15 +18,26 @@ export function signup(username, password) {
         'Content-Type': 'application/json',
       },
       body: JSON.stringify({
-        username: username.value,
-        password: password.value,
+        username,
+        password,
       }),
     })
       .then(response => response.json())
-      .then(json => dispatch({
-        type: SIGNUP_SUCCESS,
-        payload: json,
-      }))
+      .then(json => {
+        if (json.success) return json;
+
+        throw new Error(json.message);
+      })
+      .then(json => {
+        if (!json.token) throw new Error('Token has not been provided');
+
+        localStorage.setItem('token', json.token);
+
+        dispatch({
+          type: SIGNUP_SUCCESS,
+          payload: json,
+        });
+      })
       .catch(reason => dispatch({
         type: SIGNUP_FAILURE,
         payload: reason,
@@ -48,24 +58,29 @@ export function login(username, password) {
         'Content-Type': 'application/json',
       },
       body: JSON.stringify({
-        username: username.value,
-        password: password.value,
+        username,
+        password,
       }),
     })
       .then(response => response.json())
-      .then(json => dispatch({
-        type: LOGIN_SUCCESS,
-        payload: json,
-      }))
+      .then(json => {
+        if (json.success) return json;
+
+        throw new Error(json.message);
+      })
+      .then(json => {
+        if (!json.token) throw new Error('Token has not been provided');
+
+        localStorage.setItem('token', json.token);
+
+        dispatch({
+          type: LOGIN_SUCCESS,
+          payload: json,
+        });
+      })
       .catch(reason => dispatch({
         type: LOGIN_FAILURE,
         payload: reason,
       }));
-  };
-}
-
-export function logout() {
-  return (dispatch) => {
-
   };
 }
