+++ src/redux/containers/WelcomePage.js
@@ -1,18 +1,11 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { signup, login } from '../actions';
-import WelcomePage from '../components/WelcomePage';
+import { signup, login } from '../actions/index';
+import WelcomePage from '../../pages/welcomePage/WelcomePage';
 
 const mapStateToProps = state => ({
   isAuthenticated: state.auth.isAuthenticated,
 });
-
-/*
-const mapDispatchToProps = dispatch => ({
-  signup: (username, password) => dispatch(signup(username, password)),
-  login: (username, password) => dispatch(login(username, password)),
-});
-*/
 
 const mapDispatchToProps = dispatch => bindActionCreators({
   signup,
