+++ src/redux/reducers/auth.js
@@ -2,16 +2,19 @@ import {
   SIGNUP_REQUEST, SIGNUP_SUCCESS, SIGNUP_FAILURE,
   LOGIN_REQUEST, LOGIN_SUCCESS, LOGIN_FAILURE,
   LOGOUT_REQUEST, LOGOUT_SUCCESS, LOGOUT_FAILURE,
-} from '../constants';
+} from '../constants/index';
+
+const token = localStorage.getItem('token');
 
 const initialState = {
-  isAuthenticated: false,
+  isAuthenticated: !!token,
   user: {},
-  token: '',
+  token,
 };
 
 export default function auth(state = initialState, action) {
   console.log(action);
+
   switch (action.type) {
     case SIGNUP_SUCCESS:
     case LOGIN_SUCCESS:
@@ -21,10 +24,8 @@ export default function auth(state = initialState, action) {
         user: action.payload.user,
         token: action.payload.token,
       };
-
     case SIGNUP_FAILURE:
     case LOGIN_FAILURE:
-    case LOGOUT_FAILURE:
     case LOGOUT_SUCCESS:
       return {
         ...state,
