+++ src/redux/reducers/index.js
@@ -3,4 +3,4 @@ import auth from './auth';
 
 export default combineReducers({
   auth,
-})
+});
