+++ src/redux/store/index.js
@@ -1,14 +1,12 @@
 import { applyMiddleware, createStore } from 'redux';
-import rootReducer from '../reducers';
+import rootReducer from '../reducers/index';
 import thunkMiddleware from 'redux-thunk';
 
 export default function configureStore() {
   return createStore(
     rootReducer,
     applyMiddleware(
-      thunkMiddleware
-      )
-
-  )
-  
+      thunkMiddleware,
+    ),
+  );
 }
