+++ src/utils/color-from.js
@@ -6,8 +6,8 @@ import red from '@material-ui/core/colors/red';
 import green from '@material-ui/core/colors/green';
 import indigo from '@material-ui/core/colors/indigo';
 import blueGrey from '@material-ui/core/colors/blueGrey';
-import teal from '@material-ui/core/colors/teal'
-import amber from '@material-ui/core/colors/amber'
+import teal from '@material-ui/core/colors/teal';
+import amber from '@material-ui/core/colors/amber';
 
 const colors = [pink, blue, amber, orange, purple, teal, red, green, indigo, blueGrey];
